+++ forest/bonsai/simulate_gps_data.py
@@ -18,10 +18,10 @@ import ratelimit
 import requests
 from timezonefinder import TimezoneFinder
 
-from forest.constants import (ORS_API_BASE_URL, ORS_API_CALLS_PER_MINUTE,
+from ..poplar.constants.gen_constants import (ORS_API_BASE_URL, ORS_API_CALLS_PER_MINUTE,
                               OSM_OVERPASS_URL)
-from forest.jasmine.data2mobmat import great_circle_dist
-from forest.poplar.legacy.common_funcs import datetime2stamp, stamp2datetime
+from ..jasmine.data2mobmat import great_circle_dist
+from ..poplar.legacy.common_funcs import datetime2stamp, stamp2datetime
 
 R = 6.371*10**6
 ACTIVE_STATUS_LIST = range(11)
